Rename source dir to models
