+++ src/tdastro/models/basic_models.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from tdastro.sources.physical_model import SEDModel
+from tdastro.models.physical_model import SEDModel
 
 
 class ConstantSEDModel(SEDModel):
