+++ src/tdastro/models/physical_model.py
