+++ tests/tdastro/models/test_snia.py
