+++ tests/tdastro/models/test_snia_rates.py
