+++ tests/tdastro/test_simulate.py
@@ -3,9 +3,9 @@ import pytest
 from tdastro.astro_utils.passbands import PassbandGroup
 from tdastro.graph_state import GraphState
 from tdastro.math_nodes.given_sampler import GivenValueList
+from tdastro.models.basic_models import ConstantSEDModel
 from tdastro.opsim.opsim import OpSim
 from tdastro.simulate import get_time_windows, simulate_lightcurves
-from tdastro.sources.basic_models import ConstantSEDModel
 
 
 def test_get_time_windows():
